- Updated to be compatible with KDE5 file structure - Added support for clamdscan if clamd is running, otherwise clamscan is used - Changed progress bar to check for completion of background ClamAV scanning process. - Updated installation and deinstallation scripts to work with KDE5 - Scan script exits with error code under error conditions, 0 for success. - Both log files are now deleted on completion of successful scan to avoid build up of log files. - Restructured nested if conditions

+++ deinstall.sh
@@ -1,8 +1,19 @@
 #!/bin/sh
-path="$(kde4-config --path services)"
-spath="$(echo ${path%:*})"
+kde5="$(which kded5)"; 
+if [ -x "$kde5" ]; then 
+  desktop=ClamScan.desktop
+  spath="$HOME/.local/share/kservices5/"
+else 
+  desktop=ClamScan.kde4.desktop
+  path="$(kde4-config --path services)"
+  spath="$(echo ${path%:*})"
+fi
+
 rm "$spath"ServiceMenus/ClamScan/ClamScan.desktop
 rm "$spath"ServiceMenus/ClamScan/ClamScan.sh
 rm -r "$spath"ServiceMenus/ClamScan/
-path2="$(kde4-config --localprefix)share/apps/solid/actions"
-rm $path2/ClamScan_action.desktop
+
+if [ ! -x "$kde5" ]; then 
+  path2="$(kde4-config --localprefix)share/apps/solid/actions"
+  rm $path2/ClamScan_action.desktop
+fi
